% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{v_length}
\alias{v_length}
\title{Calculate vector length}
\usage{
v_length(i_vect)
}
\arguments{
\item{i_vect}{A vector}
}
\value{
The length of \code{i_vect}
}
\description{
Calculate vector length
}
\examples{
v_length(c(1,1,1,1))
}
