% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{v_normalize}
\alias{v_normalize}
\title{Normalize a vector to unit length}
\usage{
v_normalize(i_vect)
}
\arguments{
\item{i_vect}{A vector}
}
\value{
\code{i_vect} normalized to unit length
}
\description{
Normalize a vector to unit length
}
\examples{
  v_normalize(c(1,1,1,1))
}
